Added a .gitignore and started working on makig whiteboard a widget.
